Support reading Kayobe environment name from file
If using separate branches for each environment, it can be useful to
read the environment name from a file committed to each branch. Or, if
sharing the same branch for all environments, it can be useful to set
the default environment to one with lower impact, e.g. dev or staging.

If no --environment option is specified and a .environment file exists
at the root of the directory, use its content as KAYOBE_ENVIRONMENT.

Also add a missing `return 1` statement which meant a missing argument
to --environment didn't cause the script to exit with an error.

Change-Id: I22b2b466e47a3b4d96931d754a4de7544f0ada5d
(cherry picked from commit 97dcffbe67df7d56baad1945355ebfd1b45d1d41)